Fix create_pile from Game and draw_cards from Player copying cards from the deck
In both the create_pile and draw_cards method there was an issue where the cards would be copied from the deck and they wouldn't be removed, thus having the same card(s) in the pile, in the player's hand and in the deck. All that was needed to do was remove the cards from the deck after they were taken from it.