Enable user input
Modify the Player class to accept user input. The play method from the class shows the user their hand and asks them which card they want to place, and it checks whether the input is valid or not. Three more methods were added to the class which display different error messages, depending on the error that occurs.