Create HumanPlayer class
Instead of using the Player class for users and having it at the same time as a parent of the AIPlayerClass, a HumanPlayer class was made, which is a child of the Player class and is used for users. Now, the Player class only serves as a parent of both the HumanPlayer (for users) class and the AIPlayer class (for the machine).