Fix deck not being duplicated correctly
During the creation of the deck, it is duplicated in order to have more cards. However, it was shallow copied when it had to be deep copied. Because the deck was shallow copied, when a property of a card was modified, it was modified in both the original and the copy, thus causing an error where the used property of some cards was set to True before they were used.