Fix AIPlayers choosing "Wild" as the selected color after placing a wild card
After placing a wild card, an AIPlayer will choose the most common color in its hand. However, when the Wild color was the most common one, it selected it too. Now, if this happens, the AIPlayer will choose the second most common color, which will surely be a normal color.